Branch will never be taken
 - wsize >= 0 by size_t
 - data is always supplied by caller
 - formant_opts_process checks lpc_ord <= MAXORDER

+++ libformant/processing.h
@@ -24,7 +24,7 @@ int formant(int lpc_order, double s_freq, double *lpca, int *n_form,
 int w_covar(short *xx, int *m, int n, int istrt, double *y, double *alpha,
             double *r0, double preemp, window_type_t w_type);
 
-int lpc(int lpc_ord, double lpc_stabl, int wsize, short *data, double *lpca,
+int lpc(int lpc_ord, double lpc_stabl, size_t wsize, short *data, double *lpca,
         double *ar, double *lpck, double *normerr, double *rms, double preemp,
         window_type_t type);
